follower ~ 95% fixed.. can cause high cpu usage...
navmesh was the reason the follower kept crashing and locking up. all the conditions were set to loot before moving and attack before following leader or party member and by time another actions had occured, or 2 or more actions occured together (moving actions) the navmesh would get confused and stop working for the whole bot. this is due to a global table navPosition that is called from the navmesh getting confused.
the high cpu usage occuring now is due to the bot being forced to clear memory and reload the nav position every time the bot cannot find a suitable movement path